llm_router/requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0

# Training
peft>=0.7.0
trl>=0.7.0
bitsandbytes>=0.41.0

# Google AI
google-generativeai>=0.3.0

# Modal (optional, for cloud training/inference)
modal>=0.55.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Development
jupyter>=1.0.0
ipython>=8.14.0
